refactored more unit tests, and added order_guest unit tests
