+++ tests/managers/vs/vs_capacity_tests.py
