+++ tests/managers/vs/vs_tests.py
@@ -1,5 +1,5 @@
 """
-    SoftLayer.tests.managers.vs_tests
+    SoftLayer.tests.managers.vs.vs_tests
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
     :license: MIT, see LICENSE for more details.
@@ -16,8 +16,7 @@ from SoftLayer import testing
 class VSTests(testing.TestCase):
 
     def set_up(self):
-        self.vs = SoftLayer.VSManager(self.client,
-                                      SoftLayer.OrderingManager(self.client))
+        self.vs = SoftLayer.VSManager(self.client, SoftLayer.OrderingManager(self.client))
 
     def test_list_instances(self):
         results = self.vs.list_instances(hourly=True, monthly=True)
@@ -155,17 +154,6 @@ class VSTests(testing.TestCase):
                                 'reloadOperatingSystem',
                                 args=args,
                                 identifier=1)
-
-    @mock.patch('SoftLayer.managers.vs.VSManager._generate_create_dict')
-    def test_create_verify(self, create_dict):
-        create_dict.return_value = {'test': 1, 'verify': 1}
-
-        self.vs.verify_create_instance(test=1, verify=1, tags=['test', 'tags'])
-
-        create_dict.assert_called_once_with(test=1, verify=1)
-        self.assert_called_with('SoftLayer_Virtual_Guest',
-                                'generateOrderTemplate',
-                                args=({'test': 1, 'verify': 1},))
 
     @mock.patch('SoftLayer.managers.vs.VSManager._generate_create_dict')
     def test_create_instance(self, create_dict):
@@ -843,265 +831,3 @@ class VSTests(testing.TestCase):
                                 args=args,
                                 identifier=1)
 
-    def test_upgrade(self):
-        # test single upgrade
-        result = self.vs.upgrade(1, cpus=4, public=False)
-
-        self.assertEqual(result, True)
-        self.assert_called_with('SoftLayer_Product_Order', 'placeOrder')
-        call = self.calls('SoftLayer_Product_Order', 'placeOrder')[0]
-        order_container = call.args[0]
-        self.assertEqual(order_container['prices'], [{'id': 1007}])
-        self.assertEqual(order_container['virtualGuests'], [{'id': 1}])
-
-    def test_upgrade_blank(self):
-        # Now test a blank upgrade
-        result = self.vs.upgrade(1)
-
-        self.assertEqual(result, False)
-        self.assertEqual(self.calls('SoftLayer_Product_Order', 'placeOrder'),
-                         [])
-
-    def test_upgrade_full(self):
-        # Testing all parameters Upgrade
-        result = self.vs.upgrade(1,
-                                 cpus=4,
-                                 memory=2,
-                                 nic_speed=1000,
-                                 public=True)
-
-        self.assertEqual(result, True)
-        self.assert_called_with('SoftLayer_Product_Order', 'placeOrder')
-        call = self.calls('SoftLayer_Product_Order', 'placeOrder')[0]
-        order_container = call.args[0]
-        self.assertIn({'id': 1144}, order_container['prices'])
-        self.assertIn({'id': 1133}, order_container['prices'])
-        self.assertIn({'id': 1122}, order_container['prices'])
-        self.assertEqual(order_container['virtualGuests'], [{'id': 1}])
-
-    def test_upgrade_with_flavor(self):
-        # Testing Upgrade with parameter preset
-        result = self.vs.upgrade(1,
-                                 preset="M1_64X512X100",
-                                 nic_speed=1000,
-                                 public=True)
-
-        self.assertEqual(result, True)
-        self.assert_called_with('SoftLayer_Product_Order', 'placeOrder')
-        call = self.calls('SoftLayer_Product_Order', 'placeOrder')[0]
-        order_container = call.args[0]
-        self.assertEqual(799, order_container['presetId'])
-        self.assertIn({'id': 1}, order_container['virtualGuests'])
-        self.assertIn({'id': 1122}, order_container['prices'])
-        self.assertEqual(order_container['virtualGuests'], [{'id': 1}])
-
-    def test_upgrade_dedicated_host_instance(self):
-        mock = self.set_mock('SoftLayer_Virtual_Guest', 'getUpgradeItemPrices')
-        mock.return_value = fixtures.SoftLayer_Virtual_Guest.DEDICATED_GET_UPGRADE_ITEM_PRICES
-
-        # test single upgrade
-        result = self.vs.upgrade(1, cpus=4, public=False)
-
-        self.assertEqual(result, True)
-        self.assert_called_with('SoftLayer_Product_Order', 'placeOrder')
-        call = self.calls('SoftLayer_Product_Order', 'placeOrder')[0]
-        order_container = call.args[0]
-        self.assertEqual(order_container['prices'], [{'id': 115566}])
-        self.assertEqual(order_container['virtualGuests'], [{'id': 1}])
-
-    def test_get_item_id_for_upgrade(self):
-        item_id = 0
-        package_items = self.client['Product_Package'].getItems(id=46)
-        for item in package_items:
-            if ((item['prices'][0]['categories'][0]['id'] == 3)
-                    and (item.get('capacity') == '2')):
-                item_id = item['prices'][0]['id']
-                break
-        self.assertEqual(1133, item_id)
-
-    def test_get_package_items(self):
-        self.vs._get_package_items()
-        self.assert_called_with('SoftLayer_Product_Package', 'getItems')
-
-    def test_get_price_id_for_upgrade(self):
-        package_items = self.vs._get_package_items()
-
-        price_id = self.vs._get_price_id_for_upgrade(package_items=package_items,
-                                                     option='cpus',
-                                                     value='4')
-        self.assertEqual(1144, price_id)
-
-    def test_get_price_id_for_upgrade_skips_location_price(self):
-        package_items = self.vs._get_package_items()
-
-        price_id = self.vs._get_price_id_for_upgrade(package_items=package_items,
-                                                     option='cpus',
-                                                     value='55')
-        self.assertEqual(None, price_id)
-
-    def test_get_price_id_for_upgrade_finds_nic_price(self):
-        package_items = self.vs._get_package_items()
-
-        price_id = self.vs._get_price_id_for_upgrade(package_items=package_items,
-                                                     option='memory',
-                                                     value='2')
-        self.assertEqual(1133, price_id)
-
-    def test_get_price_id_for_upgrade_finds_memory_price(self):
-        package_items = self.vs._get_package_items()
-
-        price_id = self.vs._get_price_id_for_upgrade(package_items=package_items,
-                                                     option='nic_speed',
-                                                     value='1000')
-        self.assertEqual(1122, price_id)
-
-
-class VSWaitReadyGoTests(testing.TestCase):
-
-    def set_up(self):
-        self.client = mock.MagicMock()
-        self.vs = SoftLayer.VSManager(self.client)
-        self.guestObject = self.client['Virtual_Guest'].getObject
-
-    @mock.patch('SoftLayer.managers.vs.VSManager.wait_for_ready')
-    def test_wait_interface(self, ready):
-        # verify interface to wait_for_ready is intact
-        self.vs.wait_for_transaction(1, 1)
-        ready.assert_called_once_with(1, 1, delay=10, pending=True)
-
-    def test_active_not_provisioned(self):
-        # active transaction and no provision date should be false
-        self.guestObject.return_value = {'activeTransaction': {'id': 1}}
-        value = self.vs.wait_for_ready(1, 0)
-        self.assertFalse(value)
-
-    def test_active_and_provisiondate(self):
-        # active transaction and provision date should be True
-        self.guestObject.side_effect = [
-            {'activeTransaction': {'id': 1},
-             'provisionDate': 'aaa'},
-        ]
-        value = self.vs.wait_for_ready(1, 1)
-        self.assertTrue(value)
-
-    @mock.patch('time.sleep')
-    @mock.patch('time.time')
-    def test_active_provision_pending(self, _now, _sleep):
-        _now.side_effect = [0, 0, 1, 1, 2, 2]
-        # active transaction and provision date
-        # and pending should be false
-        self.guestObject.return_value = {'activeTransaction': {'id': 2}, 'provisionDate': 'aaa'}
-
-        value = self.vs.wait_for_ready(instance_id=1, limit=1, delay=1, pending=True)
-        _sleep.assert_has_calls([mock.call(0)])
-        self.assertFalse(value)
-
-    def test_reload_no_pending(self):
-        # reload complete, maintance transactions
-        self.guestObject.return_value = {
-            'activeTransaction': {'id': 2},
-            'provisionDate': 'aaa',
-            'lastOperatingSystemReload': {'id': 1},
-        }
-
-        value = self.vs.wait_for_ready(1, 1)
-        self.assertTrue(value)
-
-    @mock.patch('time.sleep')
-    @mock.patch('time.time')
-    def test_reload_pending(self, _now, _sleep):
-        _now.side_effect = [0, 0, 1, 1, 2, 2]
-        # reload complete, pending maintance transactions
-        self.guestObject.return_value = {'activeTransaction': {'id': 2},
-                                         'provisionDate': 'aaa',
-                                         'lastOperatingSystemReload': {'id': 1}}
-        value = self.vs.wait_for_ready(instance_id=1, limit=1, delay=1, pending=True)
-        _sleep.assert_has_calls([mock.call(0)])
-        self.assertFalse(value)
-
-    @mock.patch('time.sleep')
-    def test_ready_iter_once_incomplete(self, _sleep):
-        # no iteration, false
-        self.guestObject.return_value = {'activeTransaction': {'id': 1}}
-        value = self.vs.wait_for_ready(1, 0, delay=1)
-        self.assertFalse(value)
-        _sleep.assert_has_calls([mock.call(0)])
-
-    @mock.patch('time.sleep')
-    def test_iter_once_complete(self, _sleep):
-        # no iteration, true
-        self.guestObject.return_value = {'provisionDate': 'aaa'}
-        value = self.vs.wait_for_ready(1, 1, delay=1)
-        self.assertTrue(value)
-        self.assertFalse(_sleep.called)
-
-    @mock.patch('time.sleep')
-    def test_iter_four_complete(self, _sleep):
-        # test 4 iterations with positive match
-        self.guestObject.side_effect = [
-            {'activeTransaction': {'id': 1}},
-            {'activeTransaction': {'id': 1}},
-            {'activeTransaction': {'id': 1}},
-            {'provisionDate': 'aaa'},
-        ]
-
-        value = self.vs.wait_for_ready(1, 4, delay=1)
-        self.assertTrue(value)
-        _sleep.assert_has_calls([mock.call(1), mock.call(1), mock.call(1)])
-        self.guestObject.assert_has_calls([
-            mock.call(id=1, mask=mock.ANY), mock.call(id=1, mask=mock.ANY),
-            mock.call(id=1, mask=mock.ANY), mock.call(id=1, mask=mock.ANY),
-        ])
-
-    @mock.patch('time.time')
-    @mock.patch('time.sleep')
-    def test_iter_two_incomplete(self, _sleep, _time):
-        # test 2 iterations, with no matches
-        self.guestObject.side_effect = [
-            {'activeTransaction': {'id': 1}},
-            {'activeTransaction': {'id': 1}},
-            {'activeTransaction': {'id': 1}},
-            {'provisionDate': 'aaa'}
-        ]
-        # logging calls time.time as of pytest3.3, not sure if there is a better way of getting around that.
-        _time.side_effect = [0, 1, 2, 3, 4, 5, 6]
-        value = self.vs.wait_for_ready(1, 2, delay=1)
-        self.assertFalse(value)
-        _sleep.assert_has_calls([mock.call(1), mock.call(0)])
-        self.guestObject.assert_has_calls([
-            mock.call(id=1, mask=mock.ANY),
-            mock.call(id=1, mask=mock.ANY),
-        ])
-
-    @mock.patch('time.time')
-    @mock.patch('time.sleep')
-    def test_iter_20_incomplete(self, _sleep, _time):
-        """Wait for up to 20 seconds (sleeping for 10 seconds) for a server."""
-        self.guestObject.return_value = {'activeTransaction': {'id': 1}}
-        # logging calls time.time as of pytest3.3, not sure if there is a better way of getting around that.
-        _time.side_effect = [0, 0, 10, 10, 20, 20, 50, 60]
-        value = self.vs.wait_for_ready(1, 20, delay=10)
-        self.assertFalse(value)
-        self.guestObject.assert_has_calls([mock.call(id=1, mask=mock.ANY)])
-
-        _sleep.assert_has_calls([mock.call(10)])
-
-    @mock.patch('SoftLayer.decoration.sleep')
-    @mock.patch('SoftLayer.transports.FixtureTransport.__call__')
-    @mock.patch('time.time')
-    @mock.patch('time.sleep')
-    def test_exception_from_api(self, _sleep, _time, _vs, _dsleep):
-        """Tests escalating scale back when an excaption is thrown"""
-        _dsleep.return_value = False
-
-        self.guestObject.side_effect = [
-            exceptions.TransportError(104, "Its broken"),
-            {'activeTransaction': {'id': 1}},
-            {'provisionDate': 'aaa'}
-        ]
-        # logging calls time.time as of pytest3.3, not sure if there is a better way of getting around that.
-        _time.side_effect = [0, 1, 2, 3, 4]
-        value = self.vs.wait_for_ready(1, 20, delay=1)
-        _sleep.assert_called_once()
-        _dsleep.assert_called_once()
-        self.assertTrue(value)
